Create README.md
Background: 
Our client was a crypto investor who was keen on understanding the fear and greed index of the crypto market. This index is different from other stock market indexes as it is calculated by analyzing sentiments and emotions from different sources and crunching them to arrive at a number. The client wanted to determine the probability of the index reaching a certain number when there is fear or greed sentiment in the market. He was specifically interested in finding out the lowest point the index could reach when there is fear in the market, and the highest point it could reach when there is greed.

Problem: 
The dataset we received had just two columns, date and index. While it had over 1000 rows spanning from 2018 to 2021, we had a hard time understanding the complex requirements of the client. We had to go back and forth with the client to ensure we were on the same page.

Solution: 
After a lot of planning and brainstorming, we started playing around with the data. We realized that the data would be easier to work with if we divided it into class intervals of 10 each, for example, 0-9, 10-19, and so on. Since the index was set to a maximum of 100, we ended up with 10 class intervals instead of 100 numbers. We then counted the number of times the index reached a particular class interval instead of the number when the index was increasing or decreasing. This helped us build a model of the likelihood of the index reaching a particular class interval from another class interval when the index is increasing or decreasing. To accelerate the client's decision-making ability, we also did some conditional formatting.

Outcome: 
The client was extremely pleased with the solution, as it made it easier for him to analyze the data. He could now easily determine the probability of the index reaching a particular point, whether the market sentiment was one of fear or greed. With this knowledge, he was able to make more informed investment decisions.